refactor: rename to ServerSentEvents; new type FriendRequestEvent
