--- app/frontend/src/types/FriendStatusChangeEvent.ts
@@ -1,4 +1,0 @@
-export type FriendStatusChangeEvent = CustomEvent<{
-  userId: number;
-  isOnline: boolean;
-}>;
