+++ app/frontend/src/types/ServerSentEvents.ts
@@ -1,0 +1,9 @@
+export type FriendStatusChangeEvent = CustomEvent<{
+  userId: number;
+  isOnline: boolean;
+}>;
+
+export type FriendRequestEvent = CustomEvent<{
+  id: number;
+  status: "PENDING" | "ACCEPTED" | "DELETED";
+}>;
